pyproject.toml
[tool.poetry]
name = "pytest-filedata"
version = "0.4.0"
description = "easily load test data from files"
readme = 'README.md'
repository = "https://github.com/20c/pytest-filedata"
authors = [ "20C <code@20c.com>",]
license = "Apache-2.0"

classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Framework :: Pytest",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Testing",
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
decorator = ">=4"
requests-mock = ">=1"

[tool.poetry.dev-dependencies]
# Package docutils (0.21.post1) not found.
docutils = "<=0.21"

# tests
coverage = ">=5.3"
pytest = ">=6.1.2"
pytest-cov = ">=2.10.1"
tox = ">=4"
tox-gh-actions = ">2"

# linting
ruff = ">=0.1"
mypy = ">=0.950"
pre-commit = ">=2.13"
pyupgrade = ">=2.19"

# ctl
ctl = ">=1"
jinja2 = ">=3"
tmpl = ">=1"
twine = ">=3.3.0"

# docs
markdown-include = ">=0.5,<1"
mkdocs = ">=1.2.3"
pymdgen = ">=1"

[build-system]
requires = ["poetry-core>=1"]
build-backend = "poetry.core.masonry.api"
